separate buttons for viewing a serialized dependency list and an indented outline
If a user just wants to read all of the information in order from least to most complicated, now they can press one button, and no indentation is added.

If a user wants to see the unfamiliar portion of the dependency graph, now they can press the other button. It is shown in reverse order and there is indentation indicating which topics depend on which other topics.